Remove a majority of FF settings system and FF specific settings
