+++ Client/Frontend/Settings/SearchSettingsTableViewController.swift
@@ -167,14 +167,6 @@ class SearchSettingsTableViewController: UITableViewController {
             searchEnginePicker.delegate = self
             searchEnginePicker.selectedSearchEngineName = model.defaultEngine.shortName
             navigationController?.pushViewController(searchEnginePicker, animated: true)
-        } else if indexPath.item + 1 == model.orderedEngines.count {
-            let customSearchEngineForm = CustomSearchViewController()
-            customSearchEngineForm.profile = self.profile
-            customSearchEngineForm.successCallback = {
-                guard let window = self.view.window else { return }
-                SimpleToast().showAlertWithText(Strings.ThirdPartySearchEngineAdded, bottomContainer: window)
-            }
-            navigationController?.pushViewController(customSearchEngineForm, animated: true)
         }
         return nil
     }
